lab 11.1 v.completed

+++ lab11_1.cpp
@@ -1,13 +1,13 @@
 #include<iostream>
 using namespace std;
-int fibonacci(int);
+long long int fibonacci(long long int);
 int main(){
     int x;
     cin >> x;
     cout << fibonacci(x);
     return 0; 
 }
-int fibonacci(int x){
+long long int fibonacci(long long int x){
    if(x==0){
        return 0;
    }else if(x==1){
